backend/requirements.txt
# PDF Toolkit Web Application - Python Dependencies

# Core Web Framework
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1

# PDF Processing Libraries
PyPDF2==3.0.1
pdf2docx==0.5.8
PyMuPDF==1.23.8


# Image Processing
Pillow==10.1.0

# Document Generation
reportlab==4.0.7

# Word to PDF (Windows alternative - cross-platform)
docx2pdf==0.1.8

# Optional: For Word to PDF on Windows
# comtypes==1.2.0  # Uncomment if using Windows and have MS Word installed

# Development & Utilities
python-dotenv==1.0.0

# Cloud Storage - Azure Blob Storage
azure-storage-blob==12.19.0

# Production WSGI Server
gunicorn==21.2.0
